wrong-advanced-api-project-structure/advanced_api_project/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
blis==1.3.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.8
cloudpathlib==0.21.1
colorama==0.4.6
commonmark==0.9.1
confection==0.1.5
cryptography==43.0.3
cymem==2.0.11
distlib==0.3.8
Django==5.0.6
django-csp==4.0
djangorestframework==3.16.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
filelock==3.15.1
ggshield==1.42.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
googletrans==4.0.2
gspread==6.1.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.18.0
marshmallow-dataclass==8.5.14
mdurl==0.1.2
murmurhash==1.0.13
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
nltk==3.9.1
numpy==2.1.2
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==10.3.0
platformdirs==3.0.0
preshed==3.0.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygitguardian==1.24.0
Pygments==2.19.2
PyJWT==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==12.5.1
rsa==4.9
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
smart_open==7.3.0.post1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.5.1
textblob==0.19.0
thinc==8.3.6
tqdm==4.67.1
truststore==0.10.3
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vaderSentiment==3.3.2
virtualenv==20.26.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
